feat: handle pagination serverside for frontend
