+++ frontend/src/pages/contacts/index.tsx
@@ -1,14 +1,14 @@
 import { useApiContext } from '../../context/api';
 import { useEffect } from 'react';
 import Loading from '../../components/molecules/loading';
-import ContactTable from '../../components/organisms/contactTable';
+// import ContactTable from '../../components/organisms/contactTable';
 import { Page } from '../../layout/page';
-import App from '../../components/organisms/contactTable2';
+import ContactTable from '../../components/organisms/contactTable';
 
 export default function ContactsPage() {
-	const { getContacts, change, isLoading, contacts, setContacts } = useApiContext();
+	const { getAllContacts, change, isLoading, contacts, setContacts } = useApiContext();
 	useEffect(() => {
-		getContacts(1, 5).then(setContacts);
+		getAllContacts(1, 5).then(setContacts);
 	}, []);
 
 	return (
@@ -16,7 +16,7 @@ export default function ContactsPage() {
 			header={{
 				actions: [() => <p>Salam</p>, () => <h1>Hey</h1>],
 			}}>
-			{isLoading ? <Loading /> : <App />}
+			{isLoading ? <Loading /> : <ContactTable contact={[contacts]} />}
 		</Page>
 	);
 }
